+++ main.py
@@ -22,7 +22,7 @@ try:
     print('[UPDATE] Token verified. Launching...\n')
 
 except Exception:
-    promptAns = str(input('[ERROR] Token has expired or is missing. The following process will open tab in your web browser and request authorization to create a new token.\nType \'Y\' to proceed or \'N\' to cancel: '))
+    promptAns = str(input('[ERROR] Token is expired/missing. The following process may open a new tab in your web browser and request authorization to create a new token.\nType \'Y\' to proceed or \'N\' to cancel: '))
     if(promptAns.lower() == 'y' or promptAns.lower() == 'yes'):
 
         print('[UPDATE] Re-generating token. This process may take several minutes...')
